Refactored lobby to avoid circular dependency
