+++ thavalon-server/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "thavalon-server"
-version = "0.3.1"
+version = "0.4.0"
 authors = ["theadd336 <pauladdonizio@gmail.com>"]
 edition = "2018"
